Use `usize` rather than `u32` in QueryIter to avoid zero extensions
Loops using QueryIter spend time zero-extending u32 values to usize
so they can be used in address computations. There's no need to save
memory for iterator fields, since iterators are transient, so instead
just widen them to usize, which is more natural.

This drops the inner loop for the iterate_100k benchmark from this:

```
loop:  mov    %esi,%eax
       movss  (%r14,%rax,4),%xmm0
       addss  0x0(%rbp,%rax,4),%xmm0
       add    $0x1,%esi
       movss  %xmm0,0x0(%rbp,%rax,4)
       mov    0x30(%rsp),%rax
       mov    %rax,(%rsp)
       mov    0x38(%rsp),%r13
       cmp    %ecx,%esi
       je     done
       test   %rbp,%rbp
       jne    loop
done:
```

to this:

```
loop:  movss  (%r14,%rsi,4),%xmm0
       addss  0x0(%rbp,%rsi,4),%xmm0
       movss  %xmm0,0x0(%rbp,%rsi,4)
       add    $0x1,%rsi
       test   %rbp,%rbp
       je     done
       cmp    %rcx,%rsi
       jne    loop
done:
```

Resulting in a 29% speedup.

Testing if `%rbp` (one of the array pointers) is nonnull every time through
the loop still needs to be fixed -- then maybe it will vectorize.

+++ src/query.rs
@@ -431,7 +431,7 @@ impl<'q, 'w, Q: Query> IntoIterator for &'q mut QueryBorrow<'w, Q> {
 /// Iterator over the set of entities with the components in `Q`
 pub struct QueryIter<'q, 'w, Q: Query> {
     borrow: &'q mut QueryBorrow<'w, Q>,
-    archetype_index: u32,
+    archetype_index: usize,
     iter: ChunkIter<Q>,
 }
 
@@ -446,14 +446,14 @@ impl<'q, 'w, Q: Query> Iterator for QueryIter<'q, 'w, Q> {
         loop {
             match unsafe { self.iter.next() } {
                 None => {
-                    let archetype = self.borrow.archetypes.get(self.archetype_index as usize)?;
+                    let archetype = self.borrow.archetypes.get(self.archetype_index)?;
                     self.archetype_index += 1;
                     self.iter =
                         Q::Fetch::new(archetype).map_or(ChunkIter::EMPTY, |fetch| ChunkIter {
                             entities: archetype.entities(),
                             fetch,
                             position: 0,
-                            len: archetype.len(),
+                            len: archetype.len() as usize,
                         });
                     continue;
                 }
@@ -492,8 +492,8 @@ impl<'q, 'w, Q: Query> ExactSizeIterator for QueryIter<'q, 'w, Q> {
 struct ChunkIter<Q: Query> {
     entities: NonNull<u32>,
     fetch: Q::Fetch,
-    position: u32,
-    len: u32,
+    position: usize,
+    len: usize,
 }
 
 impl<Q: Query> ChunkIter<Q> {
@@ -510,8 +510,8 @@ impl<Q: Query> ChunkIter<Q> {
         if self.position == self.len {
             return None;
         }
-        let entity = self.entities.as_ptr().add(self.position as usize);
-        let item = self.fetch.get(self.position as usize);
+        let entity = self.entities.as_ptr().add(self.position);
+        let item = self.fetch.get(self.position);
         self.position += 1;
         Some((*entity, item))
     }
@@ -520,7 +520,7 @@ impl<Q: Query> ChunkIter<Q> {
 /// Batched version of `QueryIter`
 pub struct BatchedIter<'q, 'w, Q: Query> {
     borrow: &'q mut QueryBorrow<'w, Q>,
-    archetype_index: u32,
+    archetype_index: usize,
     batch_size: u32,
     batch: u32,
 }
@@ -533,7 +533,7 @@ impl<'q, 'w, Q: Query> Iterator for BatchedIter<'q, 'w, Q> {
 
     fn next(&mut self) -> Option<Self::Item> {
         loop {
-            let archetype = self.borrow.archetypes.get(self.archetype_index as usize)?;
+            let archetype = self.borrow.archetypes.get(self.archetype_index)?;
             let offset = self.batch_size * self.batch;
             if offset >= archetype.len() {
                 self.archetype_index += 1;
@@ -548,8 +548,8 @@ impl<'q, 'w, Q: Query> Iterator for BatchedIter<'q, 'w, Q> {
                     state: ChunkIter {
                         entities: archetype.entities(),
                         fetch,
-                        len: offset + self.batch_size.min(archetype.len() - offset),
-                        position: offset,
+                        len: (offset + self.batch_size.min(archetype.len() - offset)) as usize,
+                        position: offset as usize,
                     },
                 });
             } else {
